Revert "In the Iterator, change the board to a pointer."
Turns out this makes no difference at all.

On first glance it would seem that fewer copies of the data will exist,
and less time is spent on copying it over and over again. But since
there's only one iterator active at any given time, at most two boards
will exist, so there's not a lot of GC going on.

Maybe the compiler is able to spot that I don't write to either board,
so it converts it to a pointer without telling me about it.

This reverts commit d2afd7b93bde4e0fcc5161859092e009122c80d3.