Move the length parameter from Final() to Adjacent()
It was out of place in Final: positions aren't more or less final when
the path that led there is longer or shorter; if you got there, you can
presume you were able to get there.

On the other hand, it is conceivable that Adjacent can return different
results depending on the length of the path so far, for instance if the
agent has some kind of a fuel tank or something.

As an added bonus, moving the `gridPos` to the `dijkstra` package came
with an enormous performance hit, since we could no longer return a
`DeadEnd()` iterator when encountering keys for the second time.
Execution time for 2019-18-B went from .7s to 19s.
Moving the `totalCost` parameter to the `valid()` function allows me to
pretend that a previously dead end now isn't valid itself, which returns
the wait time to its previous lows.